Created a basic system for interactables. IInteractable and IInteracter interfaces. IInteractable defines an Area2D that is the interaction area, methods for when something enters and something exits that area, a method for OnInteract, and an interaction priority. IInteracter defines a CurrentInteracter. States now have CanInteract(), returning a Func<Character,IInteractable,bool>. Interactions are currently binded to E. Open the debug label to see the current interactable.
Started a room and door system. a GameRoom saves a list of its doors, and a dictionary of them grouped into directions. It recieves a signal when they are interacted, and currently just prints their direction.
Next step would be to add room connection logic in that place.
A door is an IInteractable. It has a direction and emits a signal on interaction.

Created a basic typed ObjectPool and added a few more util extensions.

Created a TestRoom to test door interactions. The room's collisions act kinda weird. Possible conclusion: segments/polygon segments are bad for collision.
But the door interactions work. Left one print North, right one prints South.